requirements.txt
asgiref==3.2.10
Django==2.1.15
django-cors-headers==3.4.0
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
djangorestframework==3.11.0
pyodbc==4.0.30
pytz==2020.1
sqlparse==0.3.1
wfastcgi==3.0.0
